+++ CHANGELOG.md
@@ -11,6 +11,8 @@
 - Fixed bug on macOS where using `with_decorations(false)` would cause `set_decorations(true)` to produce a transparent titlebar with no title.
 - Implemented `MonitorId::get_position` on macOS.
 - On macOS, `Window::get_current_monitor` now returns accurate values.
+- Added `WindowBuilderExt::with_resize_increments` to macOS.
+- **Breaking:** On X11, `WindowBuilderExt::with_resize_increments` and `WindowBuilderExt::with_base_size` now take `u32` values rather than `i32`.
 
 # Version 0.14.0 (2018-05-09)